create ts file
